Debounce changes in errors
We debounce re-rendering of the errors so that we spare a few cycles if
the user is producing too many changes.
