+++ src/components/hooks/useDebounce.ts
@@ -1,0 +1,23 @@
+import * as React from "react";
+
+// From: https://usehooks.com/useDebounce/
+export function useDebounce<T>(value: T, delay: number) {
+  const [debouncedValue, setDebouncedValue] = React.useState(value);
+
+  React.useEffect(
+    () => {
+      // Update debounced value after delay
+      const handler = setTimeout(() => {
+        setDebouncedValue(value);
+      }, delay);
+
+      return () => {
+        clearTimeout(handler);
+      };
+    },
+    // Only re-call effect if value or delay changes
+    [value, delay]
+  );
+
+  return debouncedValue;
+}
