+++ src/components/widgets/index.ts
@@ -3,3 +3,4 @@ export { OkScreen } from "./OkScreen";
 export { OkCancelScreen } from "./OkCancelScreen";
 export { SplashScreen } from "./SplashScreen";
 export { TextAreaAutoResized } from "./TextAreaAutoResized";
+export { useDebounce } from "../hooks/useDebounce";
